Struct portaudio_rs::device::DeviceInfo

pub struct DeviceInfo {
    pub name: String,
    pub host_api: HostApiIndex,
    pub max_input_channels: u32,
    pub max_output_channels: u32,
    pub default_low_input_latency: Duration,
    pub default_low_output_latency: Duration,
    pub default_high_input_latency: Duration,
    pub default_high_output_latency: Duration,
    pub default_sample_rate: f64,
}

Information for a specific device

Fields
name: String

Human readable name

host_api: HostApiIndex

Index of the host API this device belongs to

max_input_channels: u32

Maximal number of input channels that can be used

max_output_channels: u32

Maximal number of ouput channels that can be used

default_low_input_latency: Duration

Default input latency for interactive performance

default_low_output_latency: Duration

Default output latency for interactive performance

default_high_input_latency: Duration

Default input latency for robust non-interactive applications

default_high_output_latency: Duration

Default output latency for robust non-interactive applications

default_sample_rate: f64

Default sample rate