Enum portaudio_rs::PaError

#[repr(i32)]
pub enum PaError {
    NotInitialized,
    UnanticipatedHostError,
    InvalidChannelCount,
    InvalidSampleRate,
    InvalidDevice,
    InvalidFlag,
    SampleFormatNotSupported,
    BadIODeviceCombination,
    InsufficientMemory,
    BufferTooBig,
    BufferTooSmall,
    NullCallback,
    BadStreamPtr,
    TimedOut,
    InternalError,
    DeviceUnavailable,
    IncompatibleHostApiSpecificStreamInfo,
    StreamIsStopped,
    StreamIsNotStopped,
    InputOverflowed,
    OutputUnderflowed,
    HostApiNotFound,
    InvalidHostApi,
    CanNotReadFromACallbackStream,
    CanNotWriteToACallbackStream,
    CanNotReadFromAnOutputOnlyStream,
    CanNotWriteToAnInputOnlyStream,
    IncompatibleStreamHostApi,
    BadBufferPtr,
    UnknownError,
}

Enum for all possible errors given by PortAudio

The NoError value (0) is not present since the Result type can be used then.

Variants

NotInitialized

UnanticipatedHostError

InvalidChannelCount

InvalidSampleRate

InvalidDevice

InvalidFlag

SampleFormatNotSupported

BadIODeviceCombination

InsufficientMemory

BufferTooBig

BufferTooSmall

NullCallback

BadStreamPtr

TimedOut

InternalError

DeviceUnavailable

IncompatibleHostApiSpecificStreamInfo

StreamIsStopped

StreamIsNotStopped

InputOverflowed

OutputUnderflowed

HostApiNotFound

InvalidHostApi

CanNotReadFromACallbackStream

CanNotWriteToACallbackStream

CanNotReadFromAnOutputOnlyStream

CanNotWriteToAnInputOnlyStream

IncompatibleStreamHostApi

BadBufferPtr

UnknownError

Added variant for when FromPrimitive returns None

Methods

impl PaError

pub fn from_i32(num: i32) -> PaError

Get the enum value corresponding to the given i32