Struct portaudio_rs::hostapi::HostApiInfo

pub struct HostApiInfo {
    pub api_type: HostApiType,
    pub name: String,
    pub device_count: u32,
    pub default_input: Option<u32>,
    pub default_output: Option<u32>,
}

Information about a specific host API

Fields
api_type: HostApiType

The type of the API

name: String

Human-readable name of the API

device_count: u32

Number of devices this API has

default_input: Option<u32>

Default input device of the API. Is None if there is no input device available.

default_output: Option<u32>

Default output device of the API. Is None if there is no output device available.