Module portaudio_rs::stream

Contains the Stream class and associated values

Structs

Stream
An object for an PortAudio stream

StreamCallbackFlags
Flags indicating the status of the callback

StreamFlags
Flags used to control the behavior of a stream

StreamInfo
Information about the actual latency and sample rate values the stream uses

StreamParameters
Stream parameters to be used with Stream::open()

StreamTimeInfo
Time information for various stream related values

Enums

StreamCallbackResult
Allowable return values for a StreamCallback

Constants

FRAMES_PER_BUFFER_UNSPECIFIED
Argument to Stream::open() or Stream::open_default() to allow PortAudio itself determine the optimal number of frames per buffer. This number may differ each time the callback is called.

Traits

SampleType
Types that are allowed to be used as samples in a Stream

Functions

is_format_supported
Returns Ok when the StreamParameters are supported. This ignores the latency field.

Type Definitions

StreamCallback
Callback to consume, process or generate audio

StreamFinishedCallback
Callback to be fired when a StreamCallback is stopped